requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
audioread==3.0.1
av==14.4.0
beautifulsoup4==4.13.4
biopython==1.85
boto3==1.38.43
botocore==1.38.43
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
Cython==3.1.2
dataclassy==1.0.1
decorator==5.2.1
demucs==4.0.1
dnspython==2.7.0
dora_search==0.1.12
dotenv==0.9.9
dtw-python==1.5.3
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2024.6.1
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
gunicorn==21.2.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.3
httplib2==0.22.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
kneed==0.8.5
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lyricsgenius==3.6.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
Montreal_Forced_Aligner==3.2.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.0
openai-whisper==20240930
openunmix==1.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
praatio==6.2.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pymongo==4.13.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
retrying==1.3.4
rich==14.0.0
rich-click==1.8.9
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.46.2
submitit==1.5.3
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
treetable==0.2.5
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
whisper-timestamped==1.15.8
yt-dlp==2025.6.9
